# Refactor evm Instruction to be a c-like enum