rel #6744 

This allows some more type checking for the type, and also allows to put functions in a more Rust-y way.

* A small macro is used to implement `from_u8` using `match`. A small part of this marco is from `enum_derive` crate. But that crate defines other functions (like `from_u64`, etc) which we don't use.
* `INSTRUCTIONS` table is made private, and it's accessed from `Instruction::info(&self)`.
* Put `exec_stack_instruction` to `exec_instruction`. This allows Rust to type check match arms, and refuse to compile if new opcode is added, but its execution logic is not defined.