@sorpaas 2018-06-20
Here's the result got from `evmbin/bench.sh`:

This PR:

```
Output: 0x606060405200
Gas used: 29fb170
Time: 0.511926861s
^^^^ usize
Output: 0x606060405200
Gas used: 29fb170
Time: 0.519975655s
^^^^ U256
Output: 0x606060405200
Gas used: 8865053
Time: 1.749029145s
^^^^ usize
Output: 0x606060405200
Gas used: 8865053
Time: 1.787731034s
^^^^ U256
```

Master Branch:

```
Output: 0x606060405200
Gas used: 29fb170
Time: 0.573969108s
^^^^ usize
Output: 0x606060405200
Gas used: 29fb170
Time: 0.562567334s
^^^^ U256
Output: 0x606060405200
Gas used: 8865053
Time: 2.61541081s
^^^^ usize
Output: 0x606060405200
Gas used: 8865053
Time: 2.23390354s
^^^^ U256
```

Looks like somehow the conversion to enum makes it run just slightly faster..

Don't know whether this is representative enough. We can also try some benchmarks using some of the jsontests.