@tomusdrw 2018-06-25
@sorpaas Great, that's good enough for me. I think the performance tests and perhaps `evmbin/benches.rs` should be a good indicator, so I'd focus on running those.
Maybe we add some more tests to `evmbin/benches` (separate PR) and use only this in future? Will simplify comparing the changes in the future.