@5chdn 2018-06-26
Please rebase on latest master to fix CI.